[refactor][test] Move test to StreamingJobGraphGeneratorWithOperatorAttributesTest
